implement route pattern of mediainfo
